requirements.txt
asttokens==2.4.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
